Separate fetching of YUM repos and creation of files on server
