+++ tests/test_repo.py
@@ -2,217 +2,19 @@ import os
 import unittest
 from unittest import mock
 
-from argo_poem_tools.exceptions import YUMReposException
 from argo_poem_tools.repos import YUMRepos
 
-mock_data = {
-    "data": {
-        "argo-devel": {
-            "content": "[argo-devel]\n"
-                       "name=ARGO Product Repository\n"
-                       "baseurl=http://rpm-repo.argo.grnet.gr/ARGO/"
-                       "devel/centos6/\n"
-                       "gpgcheck=0\n"
-                       "enabled=1\n"
-                       "priority=99\n"
-                       "exclude=\n"
-                       "includepkgs=\n",
-            "packages": [
-                {
-                    "name": "nagios-plugins-fedcloud",
-                    "version": "0.5.0"
-                },
-                {
-                    "name": "nagios-plugins-globus",
-                    "version": "0.1.5"
-                },
-                {
-                    "name": "nagios-plugins-igtf",
-                    "version": "1.4.0"
-                },
-            ]
-        },
-        "nordugrid-updates": {
-            "content": "[nordugrid-updates]\n"
-                       "name=NorduGrid - $basearch - Updates\n"
-                       "baseurl=http://download.nordugrid.org/repos/6"
-                       "/centos/el6/$basearch/updates\n"
-                       "enabled=1\n"
-                       "gpgcheck=1\n"
-                       "gpgkey=file:///etc/pki/rpm-gpg/RPM-GPG-KEY-"
-                       "nordugrid-6\n"
-                       "priority=1\n"
-                       "exclude=ca_*\n",
-            "packages": [
-                {
-                    "name": "nordugrid-arc-nagios-plugins",
-                    "version": "2.0.0"
-                }
-            ]
-        }
-    },
-    "missing_packages": [
-        "nagios-plugins-bdii (1.0.14)",
-        "nagios-plugins-egi-notebooks (0.2.3)"]
-}
-
-mock_data_internal_metrics = {
-    "data": {
-        "argo-devel": {
-            "content": "[argo-devel]\n"
-                       "name=ARGO Product Repository\n"
-                       "baseurl=http://rpm-repo.argo.grnet.gr/ARGO/"
-                       "devel/centos6/\n"
-                       "gpgcheck=0\n"
-                       "enabled=1\n"
-                       "priority=99\n"
-                       "exclude=\n"
-                       "includepkgs=\n",
-            "packages": [
-                {
-                    "name": "argo-probe-ams-publisher",
-                    "version": "present"
-                },
-                {
-                    "name": "argo-probe-argo-tools",
-                    "version": "0.1.1"
-                },
-                {
-                    "name": "argo-probe-oidc",
-                    "version": "present"
-                },
-                {
-                    "name": "nagios-plugins-igtf",
-                    "version": "1.4.0"
-                }
-            ]
-        }
-    },
-    "missing_packages": []
-}
-
-OS_RELEASE_EL7 = \
-    b'NAME="CentOS Linux"\n' \
-    b'VERSION="7 (Core)"\n' \
-    b'ID="centos"\n' \
-    b'ID_LIKE="rhel fedora"\n' \
-    b'VERSION_ID="7"\n' \
-    b'PRETTY_NAME="CentOS Linux 7 (Core)"\n' \
-    b'ANSI_COLOR="0;31"\n' \
-    b'CPE_NAME="cpe:/o:centos:centos:7"\n' \
-    b'HOME_URL="https://www.centos.org/"\n' \
-    b'BUG_REPORT_URL="https://bugs.centos.org/"\n\n' \
-    b'CENTOS_MANTISBT_PROJECT="CentOS-7"\n' \
-    b'CENTOS_MANTISBT_PROJECT_VERSION="7"\n' \
-    b'REDHAT_SUPPORT_PRODUCT="centos"\n' \
-    b'REDHAT_SUPPORT_PRODUCT_VERSION="7"\n\n'
-
-OS_RELEASE_EL9 = \
-    b'NAME="Rocky Linux"\n' \
-    b'VERSION="9.1 (Blue Onyx)"\n' \
-    b'ID="rocky"\n' \
-    b'ID_LIKE="rhel centos fedora"\n' \
-    b'VERSION_ID="9.1"\n' \
-    b'PLATFORM_ID="platform:el9"\n' \
-    b'PRETTY_NAME="Rocky Linux 9.1 (Blue Onyx)"\n' \
-    b'ANSI_COLOR="0;32"\n' \
-    b'LOGO="fedora-logo-icon"\n' \
-    b'CPE_NAME="cpe:/o:rocky:rocky:9::baseos"\n' \
-    b'HOME_URL="https://rockylinux.org/"\n' \
-    b'BUG_REPORT_URL="https://bugs.rockylinux.org/"\n' \
-    b'ROCKY_SUPPORT_PRODUCT="Rocky-Linux-9"\n' \
-    b'ROCKY_SUPPORT_PRODUCT_VERSION="9.1"\n' \
-    b'REDHAT_SUPPORT_PRODUCT="Rocky Linux"\n' \
-    b'REDHAT_SUPPORT_PRODUCT_VERSION="9.1"\n'
-
-
-class MockResponse:
-    def __init__(self, dat, status_code):
-        self.data = dat
-        self.status_code = status_code
-        if status_code == 404:
-            self.reason = 'Not Found'
-
-        elif status_code == 403:
-            self.reason = 'Forbidden'
-
-        elif status_code == 400:
-            self.reason = 'Bad Request'
-
-        elif status_code == 500:
-            self.reason = 'Server Error'
-
-    def json(self):
-        return self.data
-
-
-def mock_request_ok(*args, **kwargs):
-    if "repos_internal" in args[0]:
-        return MockResponse(mock_data_internal_metrics, 200)
-
-    else:
-        return MockResponse(mock_data, 200)
-
-
-def mock_request_wrong_url(*args, **kwargs):
-    return MockResponse(
-        '<h1>Not Found</h1>\n'
-        '<p>The requested resource was not found on this server.</p>', 404
-    )
-
-
-def mock_request_wrong_token(*args, **kwargs):
-    return MockResponse(
-        {"detail": "Authentication credentials were not provided."}, 403
-    )
-
-
-def mock_request_wrong_profiles(*args, **kwargs):
-    return MockResponse(
-        {"detail": "You must define profile!"}, 400
-    )
-
-
-def mock_request_server_error(*args, **kwargs):
-    return MockResponse('<h1>Server Error (500)</h1>', 500)
-
-
-def mock_request_json_without_details_key(*args, **kwargs):
-    return MockResponse(
-        {'error': 'Your error message is not correct'}, 400
-    )
+from test_poem import mock_data
 
 
 class YUMReposTests(unittest.TestCase):
     def setUp(self):
         self.repos1 = YUMRepos(
-            hostname='mock.url.com',
-            token='some-token-1234',
-            profiles=['TEST_PROFILE1', 'TEST_PROFILE2'],
+            data=mock_data["data"],
             repos_path=os.getcwd()
         )
         self.repos2 = YUMRepos(
-            hostname='http://mock.url.com/',
-            token='some-token-1234',
-            profiles=['TEST_PROFILE1'],
-            repos_path=os.getcwd()
-        )
-        self.repos3 = YUMRepos(
-            hostname='https://mock.url.com/',
-            token='some-token-1234',
-            profiles=['TEST_PROFILE1', 'TEST_PROFILE2'],
-            repos_path=os.getcwd()
-        )
-        self.repos4 = YUMRepos(
-            hostname='mock.url.com',
-            token='some-token-1234',
-            profiles='',
-            repos_path=os.getcwd()
-        )
-        self.repos5 = YUMRepos(
-            hostname='mock.url.com',
-            token='some-token-1234',
-            profiles=['TEST_PROFILE1', 'TEST_PROFILE2'],
+            data=mock_data["data"],
             repos_path=os.getcwd(),
             override=False
         )
@@ -224,438 +26,8 @@ class YUMReposTests(unittest.TestCase):
         if os.path.exists('nordugrid-updates.repo'):
             os.remove('nordugrid-updates.repo')
 
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.requests.get')
-    def test_get_data_el7(self, mock_request, mock_sp):
-        mock_request.side_effect = mock_request_ok
-        mock_sp.return_value = OS_RELEASE_EL7
-        data = self.repos1.get_data()
-        mock_request.assert_called_once_with(
-            'https://mock.url.com/api/v2/repos/centos7',
-            headers={'x-api-key': 'some-token-1234',
-                     'profiles': '[TEST_PROFILE1, TEST_PROFILE2]'},
-            timeout=180
-        )
-        self.assertEqual(data, mock_data['data'])
-        self.assertEqual(
-            self.repos1.missing_packages,
-            [
-                'nagios-plugins-bdii (1.0.14)',
-                'nagios-plugins-egi-notebooks (0.2.3)'
-            ]
-        )
-
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.requests.get')
-    def test_get_data_el9(self, mock_request, mock_sp):
-        mock_request.side_effect = mock_request_ok
-        mock_sp.return_value = OS_RELEASE_EL9
-        data = self.repos1.get_data()
-        mock_request.assert_called_once_with(
-            'https://mock.url.com/api/v2/repos/rocky9',
-            headers={'x-api-key': 'some-token-1234',
-                     'profiles': '[TEST_PROFILE1, TEST_PROFILE2]'},
-            timeout=180
-        )
-        self.assertEqual(data, mock_data['data'])
-        self.assertEqual(
-            self.repos1.missing_packages,
-            [
-                'nagios-plugins-bdii (1.0.14)',
-                'nagios-plugins-egi-notebooks (0.2.3)'
-            ]
-        )
-
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.requests.get')
-    def test_get_data_including_internal_metrics(self, mock_request, mock_sp):
-        mock_request.side_effect = mock_request_ok
-        mock_sp.return_value = OS_RELEASE_EL9
-        data = self.repos1.get_data(include_internal=True)
-        self.assertEqual(mock_request.call_count, 2)
-        mock_request.assert_has_calls([
-            mock.call(
-                "https://mock.url.com/api/v2/repos/rocky9",
-                headers={'x-api-key': 'some-token-1234',
-                         'profiles': '[TEST_PROFILE1, TEST_PROFILE2]'},
-                timeout=180
-            ),
-            mock.call(
-                "https://mock.url.com/api/v2/repos_internal/rocky9",
-                headers={'x-api-key': 'some-token-1234'},
-                timeout=180
-            )
-        ], any_order=True)
-        self.assertEqual(
-            data, {
-                "argo-devel": {
-                    "content": "[argo-devel]\n"
-                               "name=ARGO Product Repository\n"
-                               "baseurl=http://rpm-repo.argo.grnet.gr/ARGO/"
-                               "devel/centos6/\n"
-                               "gpgcheck=0\n"
-                               "enabled=1\n"
-                               "priority=99\n"
-                               "exclude=\n"
-                               "includepkgs=\n",
-                    "packages": [
-                        {
-                            "name": "argo-probe-ams-publisher",
-                            "version": "present"
-                        },
-                        {
-                            "name": "argo-probe-argo-tools",
-                            "version": "0.1.1"
-                        },
-                        {
-                            "name": "argo-probe-oidc",
-                            "version": "present"
-                        },
-                        {
-                            "name": "nagios-plugins-fedcloud",
-                            "version": "0.5.0"
-                        },
-                        {
-                            "name": "nagios-plugins-globus",
-                            "version": "0.1.5"
-                        },
-                        {
-                            "name": "nagios-plugins-igtf",
-                            "version": "1.4.0"
-                        }
-                    ]
-                },
-                "nordugrid-updates": {
-                    "content": "[nordugrid-updates]\n"
-                               "name=NorduGrid - $basearch - Updates\n"
-                               "baseurl=http://download.nordugrid.org/repos/6"
-                               "/centos/el6/$basearch/updates\n"
-                               "enabled=1\n"
-                               "gpgcheck=1\n"
-                               "gpgkey=file:///etc/pki/rpm-gpg/RPM-GPG-KEY-"
-                               "nordugrid-6\n"
-                               "priority=1\n"
-                               "exclude=ca_*\n",
-                    "packages": [
-                        {
-                            "name": "nordugrid-arc-nagios-plugins",
-                            "version": "2.0.0"
-                        }
-                    ]
-                }
-            }
-        )
-        self.assertEqual(
-            self.repos1.missing_packages,
-            [
-                'nagios-plugins-bdii (1.0.14)',
-                'nagios-plugins-egi-notebooks (0.2.3)'
-            ]
-        )
-
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.requests.get')
-    def test_get_data_if_hostname_http(self, mock_request, mock_sp):
-        mock_request.side_effect = mock_request_ok
-        mock_sp.return_value = OS_RELEASE_EL9
-        data = self.repos2.get_data()
-        mock_request.assert_called_once_with(
-            'https://mock.url.com/api/v2/repos/rocky9',
-            headers={'x-api-key': 'some-token-1234',
-                     'profiles': '[TEST_PROFILE1]'},
-            timeout=180
-        )
-        self.assertEqual(data, mock_data['data'])
-        self.assertEqual(
-            self.repos2.missing_packages,
-            [
-                'nagios-plugins-bdii (1.0.14)',
-                'nagios-plugins-egi-notebooks (0.2.3)'
-            ]
-        )
-
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.requests.get')
-    def test_get_data_if_hostname_http_including_internal(
-            self, mock_request, mock_sp
-    ):
-        mock_request.side_effect = mock_request_ok
-        mock_sp.return_value = OS_RELEASE_EL9
-        data = self.repos2.get_data(include_internal=True)
-        mock_request.assert_has_calls([
-            mock.call(
-                "https://mock.url.com/api/v2/repos/rocky9",
-                headers={'x-api-key': 'some-token-1234',
-                         'profiles': '[TEST_PROFILE1]'},
-                timeout=180
-            ),
-            mock.call(
-                "https://mock.url.com/api/v2/repos_internal/rocky9",
-                headers={'x-api-key': 'some-token-1234'},
-                timeout=180
-            )
-        ], any_order=True)
-        self.assertEqual(
-            data, {
-                "argo-devel": {
-                    "content": "[argo-devel]\n"
-                               "name=ARGO Product Repository\n"
-                               "baseurl=http://rpm-repo.argo.grnet.gr/ARGO/"
-                               "devel/centos6/\n"
-                               "gpgcheck=0\n"
-                               "enabled=1\n"
-                               "priority=99\n"
-                               "exclude=\n"
-                               "includepkgs=\n",
-                    "packages": [
-                        {
-                            "name": "argo-probe-ams-publisher",
-                            "version": "present"
-                        },
-                        {
-                            "name": "argo-probe-argo-tools",
-                            "version": "0.1.1"
-                        },
-                        {
-                            "name": "argo-probe-oidc",
-                            "version": "present"
-                        },
-                        {
-                            "name": "nagios-plugins-fedcloud",
-                            "version": "0.5.0"
-                        },
-                        {
-                            "name": "nagios-plugins-globus",
-                            "version": "0.1.5"
-                        },
-                        {
-                            "name": "nagios-plugins-igtf",
-                            "version": "1.4.0"
-                        }
-                    ]
-                },
-                "nordugrid-updates": {
-                    "content": "[nordugrid-updates]\n"
-                               "name=NorduGrid - $basearch - Updates\n"
-                               "baseurl=http://download.nordugrid.org/repos/6"
-                               "/centos/el6/$basearch/updates\n"
-                               "enabled=1\n"
-                               "gpgcheck=1\n"
-                               "gpgkey=file:///etc/pki/rpm-gpg/RPM-GPG-KEY-"
-                               "nordugrid-6\n"
-                               "priority=1\n"
-                               "exclude=ca_*\n",
-                    "packages": [
-                        {
-                            "name": "nordugrid-arc-nagios-plugins",
-                            "version": "2.0.0"
-                        }
-                    ]
-                }
-            }
-        )
-        self.assertEqual(
-            self.repos2.missing_packages,
-            [
-                'nagios-plugins-bdii (1.0.14)',
-                'nagios-plugins-egi-notebooks (0.2.3)'
-            ]
-        )
-
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.requests.get')
-    def test_get_data_if_hostname_https(self, mock_request, mock_sp):
-        mock_request.side_effect = mock_request_ok
-        mock_sp.return_value = OS_RELEASE_EL9
-        data = self.repos3.get_data()
-        mock_request.assert_called_once_with(
-            'https://mock.url.com/api/v2/repos/rocky9',
-            headers={'x-api-key': 'some-token-1234',
-                     'profiles': '[TEST_PROFILE1, TEST_PROFILE2]'},
-            timeout=180
-        )
-        self.assertEqual(data, mock_data['data'])
-        self.assertEqual(
-            self.repos3.missing_packages,
-            [
-                'nagios-plugins-bdii (1.0.14)',
-                'nagios-plugins-egi-notebooks (0.2.3)'
-            ]
-        )
-
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.requests.get')
-    def test_get_data_if_hostname_https_including_internal(
-            self, mock_request, mock_sp
-    ):
-        mock_request.side_effect = mock_request_ok
-        mock_sp.return_value = OS_RELEASE_EL9
-        data = self.repos3.get_data(include_internal=True)
-        mock_request.assert_has_calls([
-            mock.call(
-                "https://mock.url.com/api/v2/repos/rocky9",
-                headers={'x-api-key': 'some-token-1234',
-                         'profiles': '[TEST_PROFILE1, TEST_PROFILE2]'},
-                timeout=180
-            ),
-            mock.call(
-                "https://mock.url.com/api/v2/repos_internal/rocky9",
-                headers={'x-api-key': 'some-token-1234'},
-                timeout=180
-            )
-        ], any_order=True)
-        self.assertEqual(
-            data, {
-                "argo-devel": {
-                    "content": "[argo-devel]\n"
-                               "name=ARGO Product Repository\n"
-                               "baseurl=http://rpm-repo.argo.grnet.gr/ARGO/"
-                               "devel/centos6/\n"
-                               "gpgcheck=0\n"
-                               "enabled=1\n"
-                               "priority=99\n"
-                               "exclude=\n"
-                               "includepkgs=\n",
-                    "packages": [
-                        {
-                            "name": "argo-probe-ams-publisher",
-                            "version": "present"
-                        },
-                        {
-                            "name": "argo-probe-argo-tools",
-                            "version": "0.1.1"
-                        },
-                        {
-                            "name": "argo-probe-oidc",
-                            "version": "present"
-                        },
-                        {
-                            "name": "nagios-plugins-fedcloud",
-                            "version": "0.5.0"
-                        },
-                        {
-                            "name": "nagios-plugins-globus",
-                            "version": "0.1.5"
-                        },
-                        {
-                            "name": "nagios-plugins-igtf",
-                            "version": "1.4.0"
-                        }
-                    ]
-                },
-                "nordugrid-updates": {
-                    "content": "[nordugrid-updates]\n"
-                               "name=NorduGrid - $basearch - Updates\n"
-                               "baseurl=http://download.nordugrid.org/repos/6"
-                               "/centos/el6/$basearch/updates\n"
-                               "enabled=1\n"
-                               "gpgcheck=1\n"
-                               "gpgkey=file:///etc/pki/rpm-gpg/RPM-GPG-KEY-"
-                               "nordugrid-6\n"
-                               "priority=1\n"
-                               "exclude=ca_*\n",
-                    "packages": [
-                        {
-                            "name": "nordugrid-arc-nagios-plugins",
-                            "version": "2.0.0"
-                        }
-                    ]
-                }
-            }
-        )
-        self.assertEqual(
-            self.repos3.missing_packages,
-            [
-                'nagios-plugins-bdii (1.0.14)',
-                'nagios-plugins-egi-notebooks (0.2.3)'
-            ]
-        )
-
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.requests.get')
-    def test_get_data_if_server_error(self, mock_request, mock_sp):
-        mock_request.side_effect = mock_request_server_error
-        mock_sp.return_value = OS_RELEASE_EL9
-        with self.assertRaises(YUMReposException) as err:
-            self.repos1.get_data()
-        self.assertEqual(
-            err.exception.__str__(),
-            "Error fetching YUM repos: 500 Server Error"
-        )
-
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.requests.get')
-    def test_get_data_if_server_error_including_internal(
-            self, mock_request, mock_sp
-    ):
-        mock_request.side_effect = mock_request_server_error
-        mock_sp.return_value = OS_RELEASE_EL9
-        with self.assertRaises(YUMReposException) as err:
-            self.repos1.get_data(include_internal=True)
-        self.assertEqual(
-            err.exception.__str__(),
-            "Error fetching YUM repos: 500 Server Error"
-        )
-
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.requests.get')
-    def test_get_data_if_wrong_url(self, mock_request, mock_sp):
-        mock_request.side_effect = mock_request_wrong_url
-        mock_sp.return_value = OS_RELEASE_EL9
-        with self.assertRaises(YUMReposException) as err:
-            self.repos1.get_data()
-        self.assertEqual(
-            err.exception.__str__(),
-            "Error fetching YUM repos: 404 Not Found"
-        )
-
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.requests.get')
-    def test_get_data_if_wrong_token(self, mock_request, mock_sp):
-        mock_request.side_effect = mock_request_wrong_token
-        mock_sp.return_value = OS_RELEASE_EL9
-        with self.assertRaises(YUMReposException) as err:
-            self.repos1.get_data()
-        self.assertEqual(
-            err.exception.__str__(),
-            "Error fetching YUM repos: 403 Forbidden: "
-            "Authentication credentials were not provided."
-        )
-
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.requests.get')
-    def test_get_data_if_no_profiles(self, mock_request, mock_sp):
-        mock_request.side_effect = mock_request_wrong_profiles
-        mock_sp.return_value = OS_RELEASE_EL9
-        with self.assertRaises(YUMReposException) as err:
-            self.repos1.get_data()
-        self.assertEqual(
-            err.exception.__str__(),
-            "Error fetching YUM repos: 400 Bad Request: "
-            "You must define profile!"
-        )
-
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.requests.get')
-    def test_get_data_if_json_without_details(
-            self, mock_request, mock_sp
-    ):
-        mock_request.side_effect = mock_request_json_without_details_key
-        mock_sp.return_value = OS_RELEASE_EL9
-        with self.assertRaises(YUMReposException) as err:
-            self.repos1.get_data()
-        self.assertEqual(
-            err.exception.__str__(),
-            "Error fetching YUM repos: 400 Bad Request"
-        )
-
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.YUMRepos.get_data')
-    def test_create_file(self, mock_get_data, mock_sp):
-        mock_get_data.return_value = mock_data["data"]
-        mock_sp.return_value = OS_RELEASE_EL9
+    def test_create_file(self):
         files = self.repos1.create_file()
-        mock_get_data.assert_called_once_with(include_internal=False)
         self.assertEqual(
             files,
             [os.path.join(os.getcwd(), 'argo-devel.repo'),
@@ -677,13 +49,8 @@ class YUMReposTests(unittest.TestCase):
             content2, mock_data['data']['nordugrid-updates']['content']
         )
 
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.YUMRepos.get_data')
-    def test_create_file_including_internal(self, mock_get_data, mock_sp):
-        mock_get_data.return_value = mock_data["data"]
-        mock_sp.return_value = OS_RELEASE_EL9
+    def test_create_file_including_internal(self):
         files = self.repos1.create_file(include_internal=True)
-        mock_get_data.assert_called_once_with(include_internal=True)
         self.assertEqual(
             files,
             [os.path.join(os.getcwd(), 'argo-devel.repo'),
@@ -708,18 +75,13 @@ class YUMReposTests(unittest.TestCase):
     @mock.patch('argo_poem_tools.repos.shutil.copyfile')
     @mock.patch('argo_poem_tools.repos.os.path.isfile')
     @mock.patch('argo_poem_tools.repos.os.makedirs')
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.YUMRepos.get_data')
     def test_do_override_file_which_already_exists(
-            self, mock_get_data, mock_sp, mock_mkdir, mock_isfile, mock_cp
+            self, mock_mkdir, mock_isfile, mock_cp
     ):
-        mock_get_data.return_value = mock_data["data"]
-        mock_sp.return_value = OS_RELEASE_EL9
         with open('argo-devel.repo', 'w') as f:
             f.write('test')
 
         files = self.repos1.create_file()
-        mock_get_data.assert_called_once_with(include_internal=False)
         self.assertFalse(mock_mkdir.called)
         self.assertFalse(mock_isfile.called)
         self.assertFalse(mock_cp.called)
@@ -747,18 +109,13 @@ class YUMReposTests(unittest.TestCase):
     @mock.patch('argo_poem_tools.repos.shutil.copyfile')
     @mock.patch('argo_poem_tools.repos.os.path.isfile')
     @mock.patch('argo_poem_tools.repos.os.makedirs')
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.YUMRepos.get_data')
     def test_do_override_file_which_already_exists_including_internal(
-            self, mock_get_data, mock_sp, mock_mkdir, mock_isfile, mock_cp
+            self, mock_mkdir, mock_isfile, mock_cp
     ):
-        mock_get_data.return_value = mock_data["data"]
-        mock_sp.return_value = OS_RELEASE_EL9
         with open('argo-devel.repo', 'w') as f:
             f.write('test')
 
         files = self.repos1.create_file(include_internal=True)
-        mock_get_data.assert_called_once_with(include_internal=True)
         self.assertFalse(mock_mkdir.called)
         self.assertFalse(mock_isfile.called)
         self.assertFalse(mock_cp.called)
@@ -786,19 +143,14 @@ class YUMReposTests(unittest.TestCase):
     @mock.patch('argo_poem_tools.repos.shutil.copyfile')
     @mock.patch('argo_poem_tools.repos.os.path.isfile')
     @mock.patch('argo_poem_tools.repos.os.makedirs')
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.YUMRepos.get_data')
     def test_do_not_override_file_which_already_exists(
-            self, mock_get_data, mock_sp, mock_mkdir, mock_isfile, mock_copy
+            self, mock_mkdir, mock_isfile, mock_copy
     ):
-        mock_get_data.return_value = mock_data["data"]
-        mock_sp.return_value = OS_RELEASE_EL9
         mock_isfile.return_value = True
         with open('argo-devel.repo', 'w') as f:
             f.write('test')
 
-        files = self.repos5.create_file()
-        mock_get_data.assert_called_once_with(include_internal=False)
+        files = self.repos2.create_file()
         self.assertEqual(mock_mkdir.call_count, 2)
         mock_mkdir.assert_called_with('/tmp' + os.getcwd(), exist_ok=True)
         file1 = os.path.join(os.getcwd(), 'argo-devel.repo')
@@ -832,19 +184,14 @@ class YUMReposTests(unittest.TestCase):
     @mock.patch('argo_poem_tools.repos.shutil.copyfile')
     @mock.patch('argo_poem_tools.repos.os.path.isfile')
     @mock.patch('argo_poem_tools.repos.os.makedirs')
-    @mock.patch('argo_poem_tools.repos.subprocess.check_output')
-    @mock.patch('argo_poem_tools.repos.YUMRepos.get_data')
     def test_do_not_override_file_which_already_exists_including_internal(
-            self, mock_get_data, mock_sp, mock_mkdir, mock_isfile, mock_copy
+            self, mock_mkdir, mock_isfile, mock_copy
     ):
-        mock_get_data.return_value = mock_data["data"]
-        mock_sp.return_value = OS_RELEASE_EL9
         mock_isfile.return_value = True
         with open('argo-devel.repo', 'w') as f:
             f.write('test')
 
-        files = self.repos5.create_file(include_internal=True)
-        mock_get_data.assert_called_once_with(include_internal=True)
+        files = self.repos2.create_file(include_internal=True)
         self.assertEqual(mock_mkdir.call_count, 2)
         mock_mkdir.assert_called_with('/tmp' + os.getcwd(), exist_ok=True)
         file1 = os.path.join(os.getcwd(), 'argo-devel.repo')
@@ -891,7 +238,7 @@ class YUMReposTests(unittest.TestCase):
         mock_isfile.return_value = True
         file1 = os.path.join(os.getcwd(), 'argo-devel.repo')
         file2 = os.path.join(os.getcwd(), 'nordugrid-updates.repo')
-        self.repos5.clean()
+        self.repos2.clean()
         self.assertEqual(mock_isdir.call_count, 1)
         mock_isdir.assert_called_with('/tmp' + os.getcwd())
         self.assertEqual(mock_ls.call_count, 1)
